telegram-bot/requirements.txt
python-telegram-bot==20.3
Flask==2.3.2
python-dotenv==1.0.0
gunicorn==20.1.0
Flask-SQLAlchemy==3.0.5
Flask-Admin==1.6.1
Flask-Login==0.6.3
Werkzeug==2.3.7
